pyproject.toml
[project]
name = "agentic-ai-assignments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.102.0",
    "openai-agents>=0.2.9",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.11",
]
